website-sale-agent/requirements.txt
litellm 
embedchain 
langchain-groq 
langchain-huggingface
stripe
colorama
python-dotenv